Beginning of Project 4
In the html I changed the date and project title and I made the script
tag external. Then in the script I added six more locations which
changed the directional functions, added to the story functions, added
separate story functions, added variables for scorekeeping, and added to
the score function.